feat(browser): alleger l'image principale en retirant Chromium/CloakBrowser (etape 2)
L'image principale n'embarque plus de navigateur : Playwright/Chromium et
CloakBrowser vivent desormais uniquement dans le runtime tracker-dashboard-browser.

- Dockerfile : suppression de `playwright install --with-deps chromium`, du bloc
  CloakBrowser et de PLAYWRIGHT_BROWSERS_PATH ; le paquet npm playwright est retire
  du runtime. curl-impersonate conserve (fast-path HTTP).
- Decouplage : le parsing des cookies + buildCookieHeader passent dans un module
  cookies.ts sans Playwright ; fetcher.ts l'importe de la. L'app principale n'a
  donc plus aucun import statique de Playwright et demarre sans navigateur.
- Les trackers mode:browser affichent un message clair tant qu'aucun runtime
  navigateur n'est configure (BROWSER_RUNTIME_URL).

Sans BROWSER_RUNTIME_URL, l'app reste pleinement fonctionnelle pour les trackers
HTTP et le fast-path curl-impersonate.
